// src/lib/types/cart.d.ts
export interface Cart {
	object: string;
	id: string;
	created_at: string;
	updated_at: string;
	deleted_at: null;
	email: string;
	billing_address_id: string | null;
	shipping_address_id: string | null;
	region_id: string;
	customer_id: string;
	payment_id: null;
	type: string;
	completed_at: null;
	payment_authorized_at: null;
	idempotency_key: null;
	context: Context;
	metadata: null;
	sales_channel_id: string;
	billing_address: Shippingaddress | null;
	discounts: any[];
	gift_cards: any[];
	items: Item[];
	payment: null;
	payment_sessions: PaymentSession[];
	payment_session: PaymentSession;
	region: Region;
	sales_channel: SalesChannel;
	shipping_address: Shippingaddress | null;
	shipping_address_id: string | null;
	shipping_methods: ShippingMethod[];
	subtotal: number;
	discount_total: number;
	item_tax_total: number;
	shipping_total: number;
	shipping_tax_total: number;
	tax_total: number;
	gift_card_total: number;
	gift_card_tax_total: number;
	total: number;
}

export interface Context {
	ip: string;
	user_agent: string;
}

export interface Item {
	id: string;
	created_at: string;
	updated_at: string;
	cart_id: string;
	order_id: null;
	swap_id: null;
	claim_order_id: null;
	original_item_id: null;
	order_edit_id: null;
	title: string;
	description: string;
	thumbnail: string;
	is_return: boolean;
	is_giftcard: boolean;
	should_merge: boolean;
	allow_discounts: boolean;
	has_shipping: boolean;
	unit_price: number;
	variant_id: string;
	quantity: number;
	fulfilled_quantity: null;
	returned_quantity: null;
	shipped_quantity: null;
	metadata: MetadataClass;
	adjustments: any[];
	tax_lines: TaxLine[];
	variant: Variant;
	subtotal: number;
	discount_total: number;
	total: number;
	original_total: number;
	original_tax_total: number;
	tax_total: number;
	raw_discount_total: number;
}

export interface MetadataClass {}

export interface TaxLine {
	rate: number;
	name: string;
	code: string;
	item_id?: string;
	shipping_method_id?: string;
}

export interface Variant {
	id: string;
	created_at: string;
	updated_at: string;
	deleted_at: null;
	title: string;
	product_id: string;
	sku: null;
	barcode: null;
	ean: null;
	upc: null;
	variant_rank: number;
	inventory_quantity: number;
	allow_backorder: boolean;
	manage_inventory: boolean;
	hs_code: null;
	origin_country: null;
	mid_code: null;
	material: null;
	weight: null;
	length: null;
	height: null;
	width: null;
	metadata: null;
	product: Product;
}
// ...
export interface SalesChannel {
	id: string;
	created_at: string;
	updated_at: string;
	deleted_at: null;
	name: string;
	type?: string;
	metadata: null;
	description?: string;
	is_disabled?: boolean;
}

export interface PaymentSession {
	id: string;
	created_at: string;
	updated_at: string;
	cart_id: string;
	provider_id: string;
	is_selected: boolean;
	is_initiated: boolean;
	status: string;
	data: PaymentSessionData;
	idempotency_key: null;
	amount: number;
	payment_authorized_at: null;
}

export interface PaymentSessionData {
	status: string;
}

export interface Region {
	id: string;
	created_at: string;
	updated_at: string;
	deleted_at: null;
	name: string;
	currency_code: string;
	tax_rate: number;
	tax_code: null;
	gift_cards_taxable: boolean;
	automatic_taxes: boolean;
	tax_provider_id: null;
	metadata: MetadataClass;
	countries: Country[];
	payment_providers: PaymentProvider[];
	tax_rates: any[];
}

export interface Country {
	id: number;
	iso_2: string;
	iso_3: string;
	num_code: number;
	name: string;
	display_name: string;
	region_id: string;
}

export interface PaymentProvider {
	id: string;
	is_installed: boolean;
}

export interface ShippingMethod {
	id: string;
	shipping_option_id: string;
	order_id: null;
	claim_order_id: null;
	cart_id: string;
	swap_id: null;
	return_id: null;
	price: number;
	data: MetadataClass;
	shipping_option: ShippingOption;
	tax_lines: TaxLine[];
	original_total: number;
	total: number;
	subtotal: number;
	original_tax_total: number;
	tax_total: number;
}

export interface ShippingOption {
	id: string;
	created_at: string;
	updated_at: string;
	deleted_at: null;
	name: string;
	region_id: string;
	profile_id: string;
	provider_id: string;
	price_type: string;
	amount: number;
	is_return: boolean;
	admin_only: boolean;
	data: ShippingOptionData;
	metadata: null;
}